dev/requirements.txt
# for prime.py:
pyserial>=3.4
